+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
-import Header from "@/components/home-address/header";
-import Main from "@/components/home-address/main";
+import Header from "@/components/home/home-header";
+import Main from "@/components/home/home-main";
 
 export default function Home() {
   return (
